requirements.txt
roboflow==1.2.1
python-dotenv==1.1.1
ultralytics==8.4.31
supervision==0.26.0